+++ .github/workflows/tests.yml
@@ -2,7 +2,6 @@ name: tests
 
 on:
   push:
-  pull_request:
 
 jobs:
   test:
