+++ docs/assets/nb_dupes.py
@@ -70,7 +70,7 @@ res_15d6 = count_dupes(15 @ P(6))
 res_8d10 = count_dupes(8 @ P(10))
 
 # %%
-from dyce.viz import plot_bar
+from dyce.viz.matplotlib import plot_bar
 
 ax = plot_bar(res_15d6, res_8d10, labels=["15d6", "8d10"])
 ax.set_title("Chances of rolling $n$ duplicates")
