+++ docs/assets/plot_d10_explode.py
@@ -35,7 +35,7 @@ def fig_callback(line_color: str) -> None:
     from matplotlib import pyplot as plt
     from matplotlib import ticker
 
-    from dyce.viz import plot_ridge
+    from dyce.viz.matplotlib import plot_ridge
 
     k_start, k_end = 3, 6  # range: [start_k..end_k)
     n_start, n_end = 5, 11  # range: [start_n..end_n)
