+++ docs/assets/plot_d4s.py
@@ -28,7 +28,7 @@ def fig_callback(line_color: str) -> None:
 
     # NOTE: Changes to this section should be propagated to docs/assets/nb_d4s.py
     # --8<-- [start:viz]
-    from dyce.viz import plot_ridge
+    from dyce.viz.matplotlib import plot_ridge
 
     ax = plot_ridge(
         h6d4p15,
