+++ docs/assets/plot_roll_and_keep.py
@@ -50,7 +50,7 @@ def fig_callback(line_color: str) -> None:
     # --8<-- [start:viz]
     from matplotlib import pyplot as plt
 
-    from dyce.viz import plot_burst
+    from dyce.viz.matplotlib import plot_burst
 
     labels1, hs1 = zip(*tuple(normal()), strict=True)
     labels2, hs2 = zip(*tuple(roll_and_keep_hs()), strict=True)
