+++ docs/assets/plot_viz_plot_line_at_most.py
@@ -17,7 +17,8 @@
 def fig_callback(line_color: str) -> None:
     # --8<-- [start:viz]
     from dyce import H
-    from dyce.viz import GraphType, plot_line
+    from dyce.viz import GraphType
+    from dyce.viz.matplotlib import plot_line
 
     ax = plot_line(
         2 @ H(10),
