+++ docs/assets/plotly_viz_plot_hbar.py
@@ -9,13 +9,13 @@
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from dyce.viz_plotly import PlotSpec
+    from dyce.viz.plotly import PlotSpec
 
 
 def fig_callback() -> "PlotSpec":
     # --8<-- [start:viz]
     from dyce import H
-    from dyce.viz_plotly import bar_spec
+    from dyce.viz.plotly import bar_spec
 
     spec = bar_spec(
         2 @ H(10),
