+++ docs/assets/plotly_viz_plot_ridge.py
@@ -16,13 +16,13 @@
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from dyce.viz_plotly import PlotSpec
+    from dyce.viz.plotly import PlotSpec
 
 
 def fig_callback() -> "PlotSpec":
     # --8<-- [start:viz]
     from dyce import H
-    from dyce.viz_plotly import ridge_spec
+    from dyce.viz.plotly import ridge_spec
 
     spec = ridge_spec(
         2 @ H(10),
