+++ docs/notes.md
@@ -53,7 +53,9 @@
 - Modernizes use of [beartype](https://github.com/beartype/beartype/) with [pytest-beartype](https://github.com/beartype/pytest-beartype)
 - Completely eliminates dependency on [`numerary`](https://github.com/beartype/numerary/) (which was flawed since conception), and instead relies on [`optype`](https://jorenham.github.io/optype/) for mathematical operator typing.
   ([`H`][dyce.H] and [`P`][dyce.P] still largely assume that outcome types won’t be mixed, but doing so will still probably work in most contexts, so FAAFO.)
-- Re-introduces low-level [Matplotlib](https://matplotlib.org/) plotting functions in [`dyce.viz`][dyce.viz]. (High-level, interactive visualization functionality still lives in [`anydyce`](https://github.com/posita/anydyce/).)
+- Re-introduces low-level [Matplotlib](https://matplotlib.org/) plotting functions in [`dyce.viz.matplotlib`][dyce.viz.matplotlib].
+- Adds portable [Plotly](https://plotly.com/) figure specifications in [`dyce.viz.plotly`][dyce.viz.plotly].
+  (High-level, interactive visualization functionality still lives in [`anydyce`](https://github.com/posita/anydyce/).)
 - Stabilizes Jupyter Lite installation
 - Defaults to collapsed installation cells in notebooks
 - Modernizes `setup.cfg` -> `pyproject.toml` (and `tox.ini`)
