+++ dyce/viz/__init__.py
@@ -13,12 +13,14 @@
 # (This does not apply to code comments.) Thank you!
 # ======================================================================================
 
+r"""Shared structures for dyce visualization backends."""
+
 import operator
 from enum import StrEnum, auto
 from itertools import accumulate
 from typing import TypeVar
 
-from .h import H
+from dyce.h import H
 
 __all__ = ("GraphType",)
